# services/backend/smart-health-platform/packages/api-contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Health & Supply Chain Resilience — API Contracts
  version: "1.0.0"
  description: >
    Authoritative single source of truth for the PHC-facing REST surface,
    offline-first delta sync engine, and authentication contracts.
    Consumed by PHC Portal, Governance Portal, and Core Backend.
servers:
  - url: http://localhost:3000
    description: Local development backend
  - url: https://api.healthresilience.gov/v1
    description: Production National DC Gateway
security:
  - bearerAuth: []
paths:
...
  /sync/push:
    post:
      summary: Upload a batch of queued offline mutations
      operationId: syncPush
      tags: [Sync]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncPushRequest'
      responses:
        '200':
          description: Batch processed (individual mutations may be accepted/duplicate/rejected/conflict)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPushResponse'
        '400':
          description: Malformed batch envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized for claimed phc_id or device binding failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sync/pull:
    get:
      summary: Download authoritative deltas since last known server sequence
      operationId: syncPull
      tags: [Sync]
      parameters:
        - name: since
          in: query
          required: true
          description: Last server_seq watermark client has applied (0 for full/initial sync)
          schema:
            type: integer
            minimum: 0
        - name: device_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
      responses:
        '200':
          description: Page of deltas since given watermark
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPullResponse'
        '410':
          description: Watermark expired / too stale — client must perform full resync
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------------
  # PHC Operational REST Endpoints (/api/v1/phc/{phcId}/*)
  # ---------------------------------------------------------------------------
...
  /api/v1/phc/{phcId}/staff:
    get:
      summary: List registered PHC staff registry and role allocations
      operationId: listStaff
      tags: [Workforce]
      parameters:
        - $ref: '#/components/parameters/phcId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StaffMember'
  /api/v1/phc/{phcId}/staff/attendance:
    post:
      summary: Record daily staff attendance
      operationId: recordAttendance
      tags: [Workforce]
      parameters:
        - $ref: '#/components/parameters/phcId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffAttendanceInput'
      responses:
        '201':
          description: Attendance recorded
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    phcId:
      name: phcId
      in: path
      required: true
      description: Target PHC facility UUID
      schema:
        type: string
        format: uuid
  schemas:
...
    SyncPushRequest:
      type: object
      required: [device_id, phc_id, mutations]
      properties:
        device_id: { type: string, format: uuid }
        phc_id: { type: string, format: uuid }
        client_clock: { type: string, format: date-time }
        mutations:
          type: array
          minItems: 1
          maxItems: 200
          items:
            $ref: '#/components/schemas/Mutation'
    Mutation:
      type: object
      required: [id, entity_type, operation, payload, local_seq, client_timestamp]
      properties:
        id:
          type: string
          format: uuid
          description: Client-generated UUID idempotency key
        entity_type:
          type: string
          enum: [billing_transaction, inventory_batch_update, resource_request, footfall_entry,
                 facility_update, staff_attendance, alert_report]
        operation:
          type: string
          enum: [create, update]
        payload:
          type: object
        local_seq:
          type: integer
        client_timestamp:
          type: string
          format: date-time
    SyncPushResponse:
      type: object
      required: [server_seq, results]
      properties:
        server_seq: { type: integer }
        results:
          type: array
          items:
            $ref: '#/components/schemas/MutationResult'
    MutationResult:
      type: object
      required: [mutation_id, status]
      properties:
        mutation_id: { type: string, format: uuid }
        status:
          type: string
          enum: [accepted, duplicate, rejected, conflict]
        server_entity_id: { type: string, format: uuid }
        error_code: { type: string }
        conflict:
          $ref: '#/components/schemas/ConflictDetail'
    ConflictDetail:
      type: object
      required: [conflict_type, server_state]
      properties:
        conflict_type:
          type: string
          enum: [stock_oversold, concurrent_absolute_edit, stale_reference]
        server_state:
          type: object
        suggested_resolution:
          type: string
          enum: [requeue_with_reduced_quantity, discard_local_edit, manual_reconciliation_required]
        reconciliation_ref: { type: string, format: uuid }
    SyncPullResponse:
      type: object
      required: [server_seq, has_more, deltas]
      properties:
        server_seq: { type: integer }
        has_more: { type: boolean }
        deltas:
          type: array
          items:
            $ref: '#/components/schemas/Delta'
    Delta:
      type: object
      required: [server_seq, entity_type, operation, entity_id, payload]
      properties:
        server_seq: { type: integer }
        entity_type:
          type: string
          enum: [redistribution_approval, request_status_change, alert, threshold_config_update, facility_config_update]
        operation:
          type: string
          enum: [upsert, delete]
        entity_id: { type: string, format: uuid }
        payload: { type: object }
        server_timestamp: { type: string, format: date-time }
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code: { type: string }
        message: { type: string }
        request_id: { type: string, format: uuid }

    # --- Domain Schemas ---
...
    StaffMember:
      type: object
      required: [id, phc_id, name, role]
      properties:
        id: { type: string, format: uuid }
        phc_id: { type: string, format: uuid }
        name: { type: string }
        role: { type: string }
        active: { type: boolean }
    StaffAttendanceInput:
      type: object
      required: [staff_id, attendance_date, status]
      properties:
        staff_id: { type: string, format: uuid }
        attendance_date: { type: string, format: date }
        status: { type: string, enum: [present, absent, on_leave] }
...